Extract point validation
